+++ src/types.sol
@@ -1,8 +1,8 @@
 // SPDX-License-Identifier: MIT
 pragma solidity ^0.8.0;
 
-struct IndexAsset{
+struct IndexAsset {
     address asset;
-    uint256 weightPercentage;
+    uint112 weightPercentage; // With 4 decimals, e.g. 50000 = 5%
     address priceFeed;
 }
